Update volume mount for raft service
Updates the sawtooth-raft volume for the raft service to mount the path
specified by the environment variable `SAWTOOTH_RAFT`; if this variable
does not exist, uses the default path `..`. This will allow the service
to be started from another Docker container with the proper path to the
sawtooth-raft directory on the host machine.

Signed-off-by: Logan Seeley <seeley@bitwise.io>