Add dynamic membership/crash fault tests to CI
Adds the dynamic membership and crash fault tolerance tests to the
Jenkinsfile. Also includes a `docker rm` command that will clean up any
residual Docker containers that were created but not removed by the
aformentioned tests; this might happen if one of the tests times out, so
it does not have a chance to clean up after itself.

Signed-off-by: Logan Seeley <seeley@bitwise.io>